purecap kernel: Update dynamic per-CPU variable support.
Dynamic per-CPU variables are managed somewhat akin to TLS variables.
The kernel defines a custom section whose contents serve as a template
for per-CPU variables.  When a CPU is initialized, a page is allocated
to hold dynamic per-CPU variables (dpcpu).  The location of a per-CPU
variable on a given CPU is computed by using the offset of the
variable in the template section (whose start adddress is DPCPU_START)
as an offset into the CPU's page.  Thus, the location of a per-CPU
variable X is:

   dpcpu + (X - DPCPU_START)

To optimize per-CPU variable accesses, hybrid and non-CHERI kernels
store the address "dpcpu - DPCPU_START" in the per-CPU 'dpcpu'
variable so that X can be accessed as:

   PCPU_GET(dpcpu) + X

In a purecap kernel, subtracting DPCPU_START moves the address far
away from the allocated page and is unrepresentable.  For purecap
kernels, undo this optimization storing the unmodified pointer in
PCPU_GET(dpcpu) and subtract DPCPU_START from X explicitly at each
access.

To support both models, introduce a DPCPU_BIAS macro that holds the
relative offset of the real dpcpu page from the per-CPU 'dpcpu'
variable.  For purecap kernels the bias is 0, but for other kernels
the bias is -DPCPU_START.

We do not set bounds on the items returned for pcpu zones.  This is
required as the zpcpu_* macros will index into the correct pcpu page
for an item.  Bounds are instead set in the zpcpu_get_* macros.

For kernel modules, export a variable to the start of the kernel's
linker set to allow it to overwrite values as the correct offset and
define DPCPU_START appropriately.

Co-authored-by: John Baldwin <jhb@FreeBSD.org>
Effort:		CHERI upstreaming
Sponsored by:	Innovate UK