fu540_prci: Rename to sifive_prci and use ocd_data for FU540 specificity
The FU740 has a very similar controller and will reuse most of the
driver. This also drops the dependency on the device-tree include for
the binding indices for two reasons. The first is that the header
doesn't namespace its contents (and nor does the FU740 one) so using
both would require separete translation units. The second is that we
don't currently have a new enough device-tree vendor import to include
the FU740's header.

Whilst here, add the missing l to gemgxlclk's name and drop the prci_
prefix from tlclk's name as we don't prefix any of the others and it's
entirely unnecessary.