nvme: Fix race condition in nvme_qpair_process_completions
Summary:
Under heavy load it is sometimes possible to hit the following panic (in
this case, on SiFive's HiFive Unmatched):

  nvme0: cpl does not map to outstanding cmd
  cdw0:00000000 sqhd:00fd sqid:0001 cid:0076 p:0 sc:00 sct:0 m:0 dnr:0
  panic: received completion for unknown cmd

This can happen if the completion's cid is read from memory before the
completion's status, as that cid read could race with the device
updating status to the current phase. Thus we must call bus_dmamap_sync
again after checking the status and before reading any other fields in
order to ensure we have acquire-like semantics.

Note that the panic message is particularly confusing. The call to
nvme_dump_completion currently passes a pointer to the real in-memory
completion (rather than our local endian-converted copy) and so, in the
case of this race, actually prints out the consistent cid, not the stale
cid that was used to get the tracker.

Reviewers: imp, kib

Subscribers: dab

Differential Revision: https://reviews.freebsd.org/D30995