purecap-kernel: acpica: Don't 1-byte pack structs with pointers
Most of the structs this macro affects don't have pointers so could be
supported just fine, but a few do so without something more granular we
have to turn the whole thing off. The fact it even exists as an option
is rather dubious anyway, the space saving is going to be minimal.