sys: Extract __bitcount* from sys/types.h to new sys/bitcount.h
Summary:
This will allow the code to be reused by the cross-build sys/types.h
wrapper in order to provide the APIs for greater compatibility. This
also provides a path towards eventually removing the definitions from
sys/types.h altogether if so desired by gradually migrating users to
including sys/bitcount.h explicitly, but that is not the primary goal
here.

Reviewers: imp, emaste

Differential Revision: https://reviews.freebsd.org/D36132