makefs: Fix issues building as a cross-tool on non-FreeBSD
Summary:
This adds missing includes, uses the standard dirent.h rather than the
BSD-specific sys/dirent.h subset (which works on macOS but not Linux)
and works around Linux's lack of st_birthtim.

This allows usr.sbin/makefs to be added to LOCAL_XTOOL_DIRS again on
macOS and Linux so that disk images can be cross-built.

Fixes:	240afd8c1fcc ("makefs: Add ZFS support")

Reviewers: markj

Subscribers: imp, delphij

Differential Revision: https://reviews.freebsd.org/D36135