Makefile.libcompat: Handle MK_FOO varying for native and compat arches
Summary:
Currently Makefile.libcompat queries a few MK_FOO variables to determine
what is being built. However, it is plausible (and indeed, downstream
in CheriBSD, this is the case) that these may vary between the native
and the compat architecture. In order to correctly determine their
values for the compat architecture, we need to defer their evaluation
until we are in the compat sub-make where src.opts.mk will give us the
right value for the compat MACHINE_ARCH.

Obtained from:	CheriBSD

Reviewers: brooks, jhb, imp, emaste

Subscribers: karels

Differential Revision: https://reviews.freebsd.org/D40570