etcupdate: Support building with BUILD_WITH_STRICT_TMPPATH enabled
In order to be able to build these various targets, we need host tools
to be available to the build, and so must first build _worldtmp to
ensure that host-symlinks is run on a skeleton worldtmp. Also, whilst we
could get away with MAKEOBJDIRPREFIX not being set for distrib-dirs, we
need it with BUILD_WITH_STRICT_TMPPATH as it determines where worldtmp
is, and thus what PATH is set to. Since every make invocation now needs
it, we can just export it up-front rather than setting it every time
(and even if it wasn't needed, it does no harm and is simplifying).