[MERGE?] sys/types.h: introduce vm_pointer_t
This type is an integer type that can hold a pointer like uintptr_t.  It
exists to denote former uses of vm_offset_t that are a pointer not an
address.  The intent is to replace it with uintptr_t or real pointer
types as appropriate over time.

Effort:		CHERI upsteraming
Sponsored by:	Innovate UK